+++ .vscode/settings.json
@@ -4,9 +4,11 @@
 	"cSpell.words": [
 		"argstest",
 		"buape",
+		"buttonping",
 		"crossbuild",
 		"crossbuildjs",
 		"disabledpagecount",
+		"itty",
 		"paginators",
 		"Rsvps",
 		"slowmode"
