+++ apps/test-worker/package.json
@@ -1,0 +1,21 @@
+{
+	"name": "test-worker",
+	"private": true,
+	"main": "src/index.ts",
+	"type": "module",
+	"scripts": {
+		"clean": "rimraf dist",
+		"deploy": "wrangler publish",
+		"lint": "eslint . --fix",
+		"tail": "wrangler tail",
+		"dev": "dotenv -- wrangler dev"
+	},
+	"dependencies": {
+		"@crossbuild/functions": "workspace:^",
+		"crossbuild": "workspace:^",
+		"wrangler": "^3.5.0"
+	},
+	"devDependencies": {
+		"@cloudflare/workers-types": "^4.20230807.0"
+	}
+}
