+++ apps/test-worker/src/components/selectMenus/ping.ts
@@ -1,0 +1,14 @@
+import { Component, CrossBuild, ReceivedInteraction } from "../../../../../packages/core/dist"
+
+export default class Cmd extends Component {
+    constructor(client: CrossBuild) {
+        super("selectping", "selectMenu", client, {
+        })
+    }
+
+    public override async run(interaction: ReceivedInteraction) {
+        await interaction.reply({
+            content: `Dropdown pong: ${interaction.selectMenuValues?.join(", ")}`
+        })
+    }
+}
