+++ packages/core/README.md
@@ -1,0 +1,8 @@
+<div align="center">
+<img width="150" height="150" src="https://cdn.buape.com/crossbuild.png" alt="CrossBuild Logo" />
+
+<h1 align="center"><b>@crossbuild/core</b></h1>
+
+This package contains the core classes and components of [CrossBuild](https://npmjs.com/package/crossbuild).
+
+</div>
