+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
-	"name": "crossbuild",
-	"description": "A powerful framework to create cross-platform Discord and Guilded bots",
+	"name": "@crossbuild/core",
+	"description": "The core classes for Crossbuild",
 	"type": "module",
 	"version": "0.3.1",
 	"contributors": [
@@ -25,7 +25,7 @@
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/buape/crossbuild.git",
-		"directory": "packages/crossbuild"
+		"directory": "packages/core"
 	},
 	"license": "MIT",
 	"bugs": {
@@ -39,14 +39,10 @@
 		"node": ">=16.17.1"
 	},
 	"dependencies": {
-		"@crossbuild/core": "workspace:^",
-		"@crossbuild/discord": "workspace:^",
 		"@crossbuild/functions": "workspace:^",
-		"@crossbuild/guilded": "workspace:^",
 		"@crossbuild/types": "workspace:^",
+		"@discordjs/collection": "^1.5.2",
 		"discord-verify": "^1.2.0",
-		"discord.js": "^14.12.1",
-		"guilded.js": "^0.23.4",
 		"itty-router": "^4.0.14"
 	}
 }
