+++ packages/core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"include": ["src/*.ts", "src/index.ts"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"rootDir": "."
+	}
+}
