+++ packages/core/typedoc.json
@@ -1,0 +1,4 @@
+{
+	"extends": ["../../typedoc.base.json"],
+	"entryPoints": ["src/index.ts"]
+}
