+++ packages/crossbuild/src/index.ts
@@ -1,11 +1,3 @@
-// export every file in the package
-export { default as Component } from "./classes/Component.js"
-export { default as ComponentHandler } from "./classes/ComponentHandler.js"
-export { default as OptionsHandler } from "./classes/OptionsHandler.js"
-export { default as ReceivedInteraction } from "./classes/ReceivedInteraction.js"
-export { default as CrossBuild } from "./classes/CrossBuild.js"
-export { default as Paginator } from "./classes/Paginator.js"
-export { default as DiscordListener } from "./listeners/Discord.js"
-export { default as GuildedListener } from "./listeners/Guilded.js"
-export * from "./typings/types.js"
-export * from "@crossbuild/types"
+export * from "@crossbuild/core"
+export * from "@crossbuild/discord"
+export * from "@crossbuild/guilded"
