+++ packages/discord/package.json
@@ -1,11 +1,11 @@
 {
-	"name": "crossbuild",
-	"description": "A powerful framework to create cross-platform Discord and Guilded bots",
+	"name": "@crossbuild/discord",
+	"description": "Discord components for CrossBuild",
 	"type": "module",
-	"version": "0.3.1",
-	"contributors": [
+	"version": "0.0.1",
+	"authors": [
 		"Shadow <shadow@buape.com>",
-		"Codeize <mrcodeize@gmail.com>"
+		"Codeize <codeize@buape.com>"
 	],
 	"scripts": {
 		"lint": "eslint . --fix",
@@ -25,7 +25,7 @@
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/buape/crossbuild.git",
-		"directory": "packages/crossbuild"
+		"directory": "packages/discord"
 	},
 	"license": "MIT",
 	"bugs": {
@@ -36,17 +36,11 @@
 		"LICENSE"
 	],
 	"engines": {
-		"node": ">=16.17.1"
+		"node": ">=18.17.0"
 	},
 	"dependencies": {
 		"@crossbuild/core": "workspace:^",
-		"@crossbuild/discord": "workspace:^",
-		"@crossbuild/functions": "workspace:^",
-		"@crossbuild/guilded": "workspace:^",
 		"@crossbuild/types": "workspace:^",
-		"discord-verify": "^1.2.0",
-		"discord.js": "^14.12.1",
-		"guilded.js": "^0.23.4",
-		"itty-router": "^4.0.14"
+		"discord.js": "^14.12.1"
 	}
 }
