+++ packages/discord/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"include": ["src/*.ts", "src/index.ts", "src/interactions/DiscordInteractionsModule.ts"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"rootDir": "."
+	}
+}
